requirements.txt
Pillow
sentence_transformers
faiss-cpu
numpy
streamlit
fastapi
